infra/cdk/requirements.txt
# AWS CDK dependencies
aws-cdk-lib>=2.120.0,<3.0.0
constructs>=10.0.0,<11.0.0
